MLIBZ-1593 Don't open two databases to check support
In order to check support for a database adapter it is needed to open the database and save an item. This change opens a database with the same name as the database used to store data for the application. This will prevent opening two databases with a size of 5mb requiring a user to click accept on a popup to allow data to be stored.